requirements.txt
joblib==1.2.0
numpy==1.23.5
pandas==2.0.0
pytorch-crf==0.7.2
scikit-learn==1.2.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
